fix: added import application to line scroll controller, changed reveal function so it targeted the project-text class
