+++ app/javascript/controllers/application.js
@@ -1,10 +1,8 @@
 import { Application } from "@hotwired/stimulus"
-import ScrollTo from 'stimulus-scroll-to'
 
 const application = Application.start()
 
 // Configure Stimulus development experience
-application.register('scroll-to', ScrollTo)
 application.debug = false
 window.Stimulus   = application
