+++ app/javascript/controllers/line_scroll_controller.js
@@ -1,9 +1,19 @@
-import { Controller } from "@hotwired/stimulus"
 
-export default class extends Controller {
-  // static targets = ["lines"]
-  connect() {
-    console.log("The 'line-scroll' controller is now loaded!")
-  }
+import { Application } from "@hotwired/stimulus"
+import ScrollTo from 'stimulus-scroll-to'
 
-}
+const application = Application.start()
+
+application.register('scroll-to', ScrollTo)
+
+let path = document.querySelector('path')
+let linesLength = path.getTotalLength()
+console.log(linesLength);
+path.style.strokeDasharray = linesLength + ' ' + linesLength;
+path.style.strokeDashoffset = linesLength;
+
+window.addEventListener('scroll', () => {
+  let scrollPercentage = (document.documentElement.scrollTop + document.body.scrollTop) / (document.documentElement.scrollHeight - document.documentElement.clientHeight);
+  let drawLength = linesLength * scrollPercentage;
+  path.style.strokeDashoffset = linesLength - drawLength;
+})
