+++ app/javascript/controllers/projects_reveal_controller.js
@@ -3,10 +3,10 @@ import { Controller } from "@hotwired/stimulus"
 export default class extends Controller {
   static targets = ["banner", "image"]
   connect() {
-    console.log("The 'projects reveal' controller is now loaded!")
+    console.log("The 'projects!!!!' controller is now loaded!")
 
     const reveal = (event) => {
-      event.currentTarget.children[3].classList.toggle("hidden")
+      event.currentTarget.querySelector('.project-text').classList.toggle("hidden");
       // event.currentTarget.children[1].classList.toggle("active")
 
     }
