+++ app/views/pages/home.html.erb
@@ -96,8 +96,8 @@
         <div class="banner-overlay">
           <div class="banner-text">Air BroomNBroom</div>
         </div>
-          <div class="project-text hidden" data-projects-reveal-target="banner">
-            <div class="project-langs">
+        <div class="project-text hidden" data-projects-reveal-target="banner">
+          <div class="project-langs">
             <ul>
               <li>Ruby</li>
               <li>Rails</li>
@@ -107,7 +107,8 @@
             </ul>
           </div>
           <div class="project-info">
-            <p>An Air BnB style website where users can hire and rent out broomsticks. Users can see their bookings on their profile page. If they are renting out products, they can also choose to accept or decline a booking request.
+            <p>
+              An Air BnB style website where users can hire and rent out broomsticks. Users can see their bookings on their profile page. If they are renting out products, they can also choose to accept or decline a booking request.
             </p>
             <a href="https://airbnb-yasbrown.herokuapp.com/" target="_blank">Visit Air Broom N Broom <i class="fa-solid fa-arrow-right"></i></a>
           </div>
